+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.4.1](https://github.com/NexusPHP/carson/compare/v1.4.0...v1.4.1) - 2026-09-12
+
 ### Changed
 
 - `pr-title-linter` and `signed-commits` success titles read `Title passes 1 rule` and `1 commit signed` instead of `Title passes all 1 rule` and `All 1 commit signed`.